friends has a confirm now

+++ src/components/friends/Friends.js
@@ -11,9 +11,14 @@ export default ({ friend }) => {
       <h3 className="friend__name">{friend.user.userName}</h3>
       <button className="btn btn-light"
         onClick={() => {
-          deleteFriend(friend)
-          // .then(() => {props.history.push("/")});
-        }}
+         const confirmDelete = window.confirm(`Are you sureyou want to delete ${friend.user.userName}? `) 
+          if (confirmDelete === true) {
+            deleteFriend(friend)
+         }
+        }
+          }
+          
+    
       >
         Remove Friend
       </button>
